created some new pages
